+++ apps/builder/app/builder/features/style-panel/style-source/index.ts
@@ -1,3 +1,3 @@
 export * from "./style-source-input";
 export * from "./style-source-badge";
-export type { ItemSource, StyleSourceError } from "./style-source";
+export type { ItemSource, StyleSourceError } from "./style-source-control";
